added a solar temperature
this is to simulate the behaviour in different conditions where solar energy is used to heat the room.
the target temperature is raised if enough solar power is available.
day one simulates a clowdy day and the condition changes every 3 minutes. this should be the worst case szenario.
day two simulates a clear day with uninterupted sun.