add forecasts to munin script
forecast_client.pl extends outside_tem.pl to include forecast temperatures.
these values are used to gain knowledge needed to decide if the heater or the AC should run (or neither)
queriend weather directly is too slow for munin, so we use gearman to keep this running and the caches hot in the background.